requirements.txt
flet==0.27.6
mysql-connector-python
